apps/api/pyproject.toml
[project]
name = "prism-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "pgvector>=0.2.5",
    "adk-prism",
    "google-adk[extensions]>=1.0.0",
    "litellm>=1.75.5",
    "google-genai>=1.0.0",
    "mcp>=1.0.0",
    "veris-ai[fastapi]>=1.20.0",
]

[project.optional-dependencies]
full = [
    "pymupdf>=1.24.0",
    "sentence-transformers>=3.0.0",
]
veris = [
    "veris-cli>=0.1.0",
]

[project.scripts]
prism-api = "prism_api.main:app_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prism_api"]
